Change views to follow mediator pattern
